requirements.txt
Pillow>=9.0.0
mediapipe>=0.10.0
numpy>=1.21.0
pillow-heif>=0.13.0
torch>=2.0.0
openai